requirements.txt
requests
pandas
python-dotenv
faker